WLTS-30-Extra-Stuff I added an extra test that has multiple unique values and updated the verify to use the GET dictionary
