+++ src/functionalTest/java/uk/gov/hmcts/reform/translate/custom/value/CustomValueKey.java
@@ -7,8 +7,7 @@ public enum CustomValueKey {
     UNIQUE_TRANSLATION_WITH_ENGLISH_AND_WELSH("UniqueTranslationWithEnglishAndWelsh"),
     UNIQUE_TRANSLATION_WITH_ONLY_ENGLISH("UniqueTranslationWithOnlyEnglish"),
     CONTAINS_DICTIONARY_TRANSLATIONS("containsDictionaryTranslations"),
-    GET_VERIFY_PHRASES_FROM_PARENT_CONTEXT("getVerifyPhrasesFromParentContext"),
-    GET_VERIFY_TRANSLATION_RESPONSE_FROM_PARENT_CONTEXT("getVerifyTranslationResponseFromParentContext"),
+    CONTAINS_UNIQUE_PHRASES_PARENT_CONTEXT("containsUniquePhrasesParentContext"),
     S_003_1("S-003.1"),
     DEFAULT_KEY("DefaultKey");
