+++ src/functionalTest/java/uk/gov/hmcts/reform/translate/custom/value/UniqueTranslationWithEnglishAndWelshEvaluator.java
@@ -3,6 +3,8 @@ package uk.gov.hmcts.reform.translate.custom.value;
 import uk.gov.hmcts.befta.player.BackEndFunctionalTestScenarioContext;
 import uk.gov.hmcts.reform.translate.StringGenerator;
 
+import java.sql.SQLOutput;
+import java.util.HashMap;
 import java.util.Map;
 
 public class UniqueTranslationWithEnglishAndWelshEvaluator implements CustomValueEvaluator {
@@ -14,8 +16,15 @@ public class UniqueTranslationWithEnglishAndWelshEvaluator implements CustomValu
 
     @Override
     public Object calculate(BackEndFunctionalTestScenarioContext scenarioContext, Object key) {
-        String englishPhrase = StringGenerator.generate();
-        final String welshPhrase = englishPhrase + "-WELSH";
-        return Map.of(englishPhrase, welshPhrase);
+        final String expectedValueStr = key.toString().replace("UniqueTranslationWithEnglishAndWelsh ", "");
+        System.out.println("sdansdjaksjdkajs"+expectedValueStr);
+        int uniquePhrasesAmount = Integer.parseInt(expectedValueStr);
+        Map<String, String> map = new HashMap<>();
+        for (int i = 0; i<uniquePhrasesAmount; i++) {
+            String englishPhrase = StringGenerator.generate();
+            final String welshPhrase = englishPhrase + "-WELSH";
+            map.put(englishPhrase, welshPhrase);
+        }
+        return map;
     }
 }
