+++ src/functionalTest/resources/features/F-004 - Put Dictionary /F-004_Verify.td.json
@@ -1,31 +1,25 @@
 {
   "_guid_": "F-004_Verify",
-  "_extends_": "Translation_Cy_Base",
-
+  "_extends_": "Get_Dictionary_Base",
+  "title": "must return a 200 OK containing all dictionary entries",
   "specs": [
     "to verify translations"
-
   ],
+
   "users": {
     "invokingUser": {
       "_extends_": "ManageTranslationsUser"
     }
   },
+
   "request": {
-    "_extends_": "Common_Request",
-    "headers": {
-      "Accept": "application/json"
-    },
-    "body": {
-      "phrases": "${[scenarioContext][customValues][getVerifyPhrasesFromParentContext]}"
-    }
+    "_extends_": "Common_Request"
   },
+
   "expectedResponse": {
     "_extends_": "Common_200_Response",
-    "responseCode": 200,
-    "responseMessage": "OK",
     "body": {
-      "translations": "${[scenarioContext][customValues][getVerifyTranslationResponseFromParentContext]}"
+      "translations": "${[scenarioContext][customValues][containsUniquePhrasesParentContext]}"
     }
   }
 }
